Got rid of lots of warnings. Switch default environment to v2, though added v1.
