+++ src/pydantic_yaml/dumper.py
@@ -11,6 +11,7 @@ from io import StringIO, IOBase
 from pathlib import Path
 from typing import Any, Union
 
+import pydantic
 from ruamel.yaml import YAML
 from pydantic import BaseModel
 
@@ -37,7 +38,10 @@ def _write_yaml_model(stream: IOBase, model: BaseModel, **kwargs) -> None:
         Keyword arguments to pass `model.json()`. FIXME: Add explicit arguments.
     """
     model = _chk_model(model)
-    json_val = model.json(**kwargs)
+    if pydantic.__version__ < "2":
+        json_val = model.json(**kwargs)
+    else:
+        json_val = model.model_dump_json(**kwargs)
     val = json.loads(json_val)
     writer = YAML(typ="safe", pure=True)
     # TODO: Configure writer
