+++ src/pydantic_yaml/examples/v2/__init__.py
@@ -1,0 +1,1 @@
+"""Examples for Pydantic v2."""
