+++ src/test/test_basic.py
@@ -8,6 +8,7 @@ import pytest
 from pydantic import BaseModel
 
 from pydantic_yaml import parse_yaml_file_as, parse_yaml_raw_as, to_yaml_str, to_yaml_file
+
 from pydantic_yaml.examples.base_models import (
     A,
     B,
